+++ DX11Renderer/MatrixAnimation/imgui.ini
@@ -3,6 +3,6 @@ Pos=60,60
 Size=400,400
 
 [Window][Scene Control]
-Pos=1178,0
-Size=422,1200
+Pos=1161,0
+Size=439,1200
