+++ DX11Renderer/shared/shaders/NormalVertexShader.hlsl
