+++ DX11Renderer/shared/src/D3DUtils.cpp
@@ -1,4 +1,4 @@
-#include "D3DUtils.h"
+#include "include/D3DUtils.h"
 
 #define STB_IMAGE_IMPLEMENTATION
 #include "stb_image.h"
