+++ DX11Renderer/shared/src/include/Vertex.h
