+++ DX11Renderer/shared/src/include/WinApp.h
